Update dot character matching for all api regexes
