+++ adapter/internal/discovery/xds/semanticVersioning.go
@@ -17,12 +17,130 @@
 package xds
 
 import (
+	"errors"
+	"strconv"
 	"strings"
 
 	routev3 "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
 	envoy_type_matcherv3 "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
+	logger "github.com/wso2/product-microgateway/adapter/internal/loggers"
 	mgw "github.com/wso2/product-microgateway/adapter/internal/oasparser/model"
 )
+
+// GetVersionMatchRegex returns the regex to match the full version string
+func GetVersionMatchRegex(version string) string {
+	// Match "." character in the version by replacing it with "\\."
+	return strings.ReplaceAll(version, ".", "\\.")
+}
+
+// ValidateAndGetVersionComponents validates version string and extracts version components
+func ValidateAndGetVersionComponents(version string, apiName string) (*SemVersion, error) {
+	versionComponents := strings.Split(version, ".")
+
+	// If the versionComponents length is less than 2, return error
+	if len(versionComponents) < 2 {
+		logger.LoggerXds.Errorf("API version validation failed for API: %v. API Version: %v", apiName, version)
+		errMessage := "Invalid version: " + version + " for API: " + apiName +
+			". API version should be in the format x.y.z where x,y,z are non-negative integers"
+		return nil, errors.New(errMessage)
+	}
+
+	majorVersion, majorVersionConvErr := strconv.Atoi(versionComponents[0])
+	minorVersion, minorVersionConvErr := strconv.Atoi(versionComponents[1])
+	if majorVersionConvErr != nil || majorVersion < 0 {
+		logger.LoggerXds.Errorf("API major version should be a non-negative integer in API: %v. API Version: %v", apiName, version, majorVersionConvErr)
+		return nil, errors.New("Invalid version format")
+	}
+
+	if minorVersionConvErr != nil || minorVersion < 0 {
+		logger.LoggerXds.Errorf("API minor version should be a non-negative integer in API: %v. API Version: %v", apiName, version, minorVersionConvErr)
+		return nil, errors.New("Invalid version format")
+	}
+
+	if len(versionComponents) == 2 {
+		return &SemVersion{
+			Version: version,
+			Major:   majorVersion,
+			Minor:   minorVersion,
+			Patch:   nil,
+		}, nil
+	}
+
+	patchVersion, patchVersionConvErr := strconv.Atoi(versionComponents[2])
+	if patchVersionConvErr != nil {
+		logger.LoggerXds.Errorf("API patch version should be an integer in API: %v. API Version: %v", apiName, version, patchVersionConvErr)
+		return nil, errors.New("Invalid version format")
+	}
+	return &SemVersion{
+		Version: version,
+		Major:   majorVersion,
+		Minor:   minorVersion,
+		Patch:   &patchVersion,
+	}, nil
+}
+
+// GetMajorMinorVersionRangeRegex generates major and minor version compatible range regex for the given version
+func GetMajorMinorVersionRangeRegex(semVersion SemVersion) string {
+	versionRegexString := GetVersionMatchRegex(semVersion.Version)
+	majorVersion := strconv.Itoa(semVersion.Major)
+	minorVersion := strconv.Itoa(semVersion.Minor)
+	if semVersion.Patch == nil {
+		return "(v" + majorVersion + "(\\." + minorVersion + ")?|" + versionRegexString + ")"
+	}
+	patchVersion := strconv.Itoa(*semVersion.Patch)
+	return "(v" + majorVersion + "(\\." + minorVersion + "(\\." + patchVersion + ")?)?|" + versionRegexString + ")"
+}
+
+// GetMinorVersionRangeRegex generates minor version compatible range regex for the given version
+func GetMinorVersionRangeRegex(semVersion SemVersion) string {
+	versionRegexString := GetVersionMatchRegex(semVersion.Version)
+	if semVersion.Patch == nil {
+		return "(v" + versionRegexString + "|" + versionRegexString + ")"
+	}
+	majorVersion := strconv.Itoa(semVersion.Major)
+	minorVersion := strconv.Itoa(semVersion.Minor)
+	patchVersion := strconv.Itoa(*semVersion.Patch)
+	return "(v" + majorVersion + "\\." + minorVersion + "(\\." + patchVersion + ")?|" + versionRegexString + ")"
+}
+
+// GetMajorVersionRange generates major version range for the given version
+func GetMajorVersionRange(semVersion SemVersion) string {
+	return "v" + strconv.Itoa(semVersion.Major)
+}
+
+// GetMinorVersionRange generates minor version range for the given version
+func GetMinorVersionRange(semVersion SemVersion) string {
+	return "v" + strconv.Itoa(semVersion.Major) + "." + strconv.Itoa(semVersion.Minor)
+}
+
+// CompareSemanticVersions compares two semantic versions and returns true
+// if `version` is greater or equal than `baseVersion`
+func CompareSemanticVersions(baseVersion, version SemVersion) bool {
+	if baseVersion.Major < version.Major {
+		return true
+	} else if baseVersion.Major > version.Major {
+		return false
+	} else {
+		if baseVersion.Minor < version.Minor {
+			return true
+		} else if baseVersion.Minor > version.Minor {
+			return false
+		} else {
+			if baseVersion.Patch != nil && version.Patch != nil {
+				if *baseVersion.Patch < *version.Patch {
+					return true
+				} else if *baseVersion.Patch > *version.Patch {
+					return false
+				}
+			} else if baseVersion.Patch == nil && version.Patch != nil {
+				return true
+			} else if baseVersion.Patch != nil && version.Patch == nil {
+				return false
+			}
+		}
+	}
+	return true
+}
 
 func updateRoutingRulesOnAPIUpdate(organizationID, apiIdentifier, apiName, apiVersion, vHost string) {
 	apiSemVersion, err := ValidateAndGetVersionComponents(apiVersion, apiName)
@@ -54,11 +172,13 @@ func updateRoutingRulesOnAPIUpdate(organizationID, apiIdentifier, apiName, apiVe
 					for _, route := range orgIDOpenAPIRoutesMap[organizationID][apiUUID] {
 						regex := route.GetMatch().GetSafeRegex().GetRegex()
 						regexRewritePattern := route.GetRoute().GetRegexRewrite().GetPattern().GetRegex()
+						existingMinorRangeLatestVersionRegex := GetVersionMatchRegex(existingMinorRangeLatestSemVersion.Version)
+						existingMajorRangeLatestVersionRegex := GetVersionMatchRegex(existingMajorRangeLatestSemVersion.Version)
 						if isMinorRangeRegexAvailable && swagger.GetVersion() == existingMinorRangeLatestSemVersion.Version && isLatestMinorVersion {
-							regex = strings.Replace(regex, GetMinorVersionRangeRegex(existingMinorRangeLatestSemVersion), existingMinorRangeLatestSemVersion.Version, 1)
-							regex = strings.Replace(regex, GetMajorMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), existingMajorRangeLatestSemVersion.Version, 1)
-							regexRewritePattern = strings.Replace(regexRewritePattern, GetMinorVersionRangeRegex(existingMinorRangeLatestSemVersion), existingMinorRangeLatestSemVersion.Version, 1)
-							regexRewritePattern = strings.Replace(regexRewritePattern, GetMajorMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), existingMajorRangeLatestSemVersion.Version, 1)
+							regex = strings.Replace(regex, GetMinorVersionRangeRegex(existingMinorRangeLatestSemVersion), existingMinorRangeLatestVersionRegex, 1)
+							regex = strings.Replace(regex, GetMajorMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), existingMajorRangeLatestVersionRegex, 1)
+							regexRewritePattern = strings.Replace(regexRewritePattern, GetMinorVersionRangeRegex(existingMinorRangeLatestSemVersion), existingMinorRangeLatestVersionRegex, 1)
+							regexRewritePattern = strings.Replace(regexRewritePattern, GetMajorMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), existingMajorRangeLatestVersionRegex, 1)
 						}
 						if isMajorRangeRegexAvailable && swagger.GetVersion() == existingMajorRangeLatestSemVersion.Version && isLatestMajorVersion {
 							regex = strings.Replace(regex, GetMajorMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), GetMinorVersionRangeRegex(existingMajorRangeLatestSemVersion), 1)
@@ -107,12 +227,13 @@ func updateRoutingRulesOnAPIUpdate(organizationID, apiIdentifier, apiName, apiVe
 		for _, route := range orgIDOpenAPIRoutesMap[organizationID][apiIdentifier] {
 			regex := route.GetMatch().GetSafeRegex().GetRegex()
 			regexRewritePattern := route.GetRoute().GetRegexRewrite().GetPattern().GetRegex()
+			apiVersionRegex := GetVersionMatchRegex(apiVersion)
 			if isLatestMajorVersion {
-				regex = strings.Replace(regex, apiVersion, GetMajorMinorVersionRangeRegex(*apiSemVersion), 1)
-				regexRewritePattern = strings.Replace(regexRewritePattern, apiVersion, GetMajorMinorVersionRangeRegex(*apiSemVersion), 1)
+				regex = strings.Replace(regex, apiVersionRegex, GetMajorMinorVersionRangeRegex(*apiSemVersion), 1)
+				regexRewritePattern = strings.Replace(regexRewritePattern, apiVersionRegex, GetMajorMinorVersionRangeRegex(*apiSemVersion), 1)
 			} else if isLatestMinorVersion {
-				regex = strings.Replace(regex, apiVersion, GetMinorVersionRangeRegex(*apiSemVersion), 1)
-				regexRewritePattern = strings.Replace(regexRewritePattern, apiVersion, GetMinorVersionRangeRegex(*apiSemVersion), 1)
+				regex = strings.Replace(regex, apiVersionRegex, GetMinorVersionRangeRegex(*apiSemVersion), 1)
+				regexRewritePattern = strings.Replace(regexRewritePattern, apiVersionRegex, GetMinorVersionRangeRegex(*apiSemVersion), 1)
 			}
 			pathSpecifier := &routev3.RouteMatch_SafeRegex{
 				SafeRegex: &envoy_type_matcherv3.RegexMatcher{
@@ -170,28 +291,29 @@ func updateRoutingRulesOnAPIDelete(organizationID, apiIdentifier string, api mgw
 				for _, route := range orgIDOpenAPIRoutesMap[organizationID][newLatestMajorRangeApiUuid] {
 					regex := route.GetMatch().GetSafeRegex().GetRegex()
 					regexRewritePattern := route.GetRoute().GetRegexRewrite().GetPattern().GetRegex()
+					newLatestMajorRangeApiVersionRegex := GetVersionMatchRegex(newLatestMajorRangeApi.Version)
 					// Remove any available minor version range regexes and apply the minor range regex
 					regex = strings.Replace(
 						regex,
 						GetMinorVersionRangeRegex(*newLatestMajorRangeApi),
-						newLatestMajorRangeApi.Version,
+						newLatestMajorRangeApiVersionRegex,
 						1,
 					)
 					regexRewritePattern = strings.Replace(
 						regexRewritePattern,
 						GetMinorVersionRangeRegex(*newLatestMajorRangeApi),
-						newLatestMajorRangeApi.Version,
+						newLatestMajorRangeApiVersionRegex,
 						1,
 					)
 					regex = strings.Replace(
 						regex,
-						newLatestMajorRangeApi.Version,
+						newLatestMajorRangeApiVersionRegex,
 						GetMajorMinorVersionRangeRegex(*newLatestMajorRangeApi),
 						1,
 					)
 					regexRewritePattern = strings.Replace(
 						regexRewritePattern,
-						newLatestMajorRangeApi.Version,
+						newLatestMajorRangeApiVersionRegex,
 						GetMajorMinorVersionRangeRegex(*newLatestMajorRangeApi),
 						1,
 					)
@@ -240,9 +362,10 @@ func updateRoutingRulesOnAPIDelete(organizationID, apiIdentifier string, api mgw
 				orgIDLatestAPIVersionMap[organizationID][apiRangeIdentifier][minorVersionRange] = *newLatestMinorRangeApi
 				for _, route := range orgIDOpenAPIRoutesMap[organizationID][newLatestMinorRangeApiUuid] {
 					regex := route.GetMatch().GetSafeRegex().GetRegex()
+					newLatestMinorRangeApiVersionRegex := GetVersionMatchRegex(newLatestMinorRangeApi.Version)
 					regex = strings.Replace(
 						regex,
-						newLatestMinorRangeApi.Version,
+						newLatestMinorRangeApiVersionRegex,
 						GetMinorVersionRangeRegex(*newLatestMinorRangeApi),
 						1,
 					)
@@ -254,7 +377,7 @@ func updateRoutingRulesOnAPIDelete(organizationID, apiIdentifier string, api mgw
 					regexRewritePattern := route.GetRoute().GetRegexRewrite().GetPattern().GetRegex()
 					regexRewritePattern = strings.Replace(
 						regexRewritePattern,
-						newLatestMinorRangeApi.Version,
+						newLatestMinorRangeApiVersionRegex,
 						GetMinorVersionRangeRegex(*newLatestMinorRangeApi),
 						1,
 					)
