+++ adapter/internal/discovery/xds/server_utils.go
@@ -17,10 +17,6 @@
 package xds
 
 import (
-	"errors"
-	"strconv"
-	"strings"
-
 	logger "github.com/wso2/product-microgateway/adapter/internal/loggers"
 )
 
@@ -84,113 +80,4 @@ func updateVhostInternalMaps(uuid, name, version, vHost string, gwEnvs []string)
 		envToVhostMap[env] = vHost
 	}
 	apiUUIDToGatewayToVhosts[uuid] = envToVhostMap
-}
-
-// ValidateAndGetVersionComponents validates version string and extracts version components
-func ValidateAndGetVersionComponents(version string, apiName string) (*SemVersion, error) {
-	versionComponents := strings.Split(version, ".")
-
-	// If the versionComponents length is less than 2, return error
-	if len(versionComponents) < 2 {
-		logger.LoggerXds.Errorf("API version validation failed for API: %v. API Version: %v", apiName, version)
-		errMessage := "Invalid version: " + version + " for API: " + apiName +
-			". API version should be in the format x.y.z where x,y,z are non-negative integers"
-		return nil, errors.New(errMessage)
-	}
-
-	majorVersion, majorVersionConvErr := strconv.Atoi(versionComponents[0])
-	minorVersion, minorVersionConvErr := strconv.Atoi(versionComponents[1])
-	if majorVersionConvErr != nil || majorVersion < 0 {
-		logger.LoggerXds.Errorf("API major version should be a non-negative integer in API: %v. API Version: %v", apiName, version, majorVersionConvErr)
-		return nil, errors.New("Invalid version format")
-	}
-
-	if minorVersionConvErr != nil || minorVersion < 0 {
-		logger.LoggerXds.Errorf("API minor version should be a non-negative integer in API: %v. API Version: %v", apiName, version, minorVersionConvErr)
-		return nil, errors.New("Invalid version format")
-	}
-
-	if len(versionComponents) == 2 {
-		return &SemVersion{
-			Version: version,
-			Major:   majorVersion,
-			Minor:   minorVersion,
-			Patch:   nil,
-		}, nil
-	}
-
-	patchVersion, patchVersionConvErr := strconv.Atoi(versionComponents[2])
-	if patchVersionConvErr != nil {
-		logger.LoggerXds.Errorf("API patch version should be an integer in API: %v. API Version: %v", apiName, version, patchVersionConvErr)
-		return nil, errors.New("Invalid version format")
-	}
-	return &SemVersion{
-		Version: version,
-		Major:   majorVersion,
-		Minor:   minorVersion,
-		Patch:   &patchVersion,
-	}, nil
-}
-
-// GetMajorMinorVersionRangeRegex generates major and minor version compatible range regex for the given version
-func GetMajorMinorVersionRangeRegex(semVersion SemVersion) string {
-	versionRegexString := strings.ReplaceAll(semVersion.Version, ".", "\\.")
-	majorVersion := strconv.Itoa(semVersion.Major)
-	minorVersion := strconv.Itoa(semVersion.Minor)
-	if semVersion.Patch == nil {
-		return "(v" + majorVersion + "(\\." + minorVersion + ")?|" + versionRegexString + ")"
-	}
-	patchVersion := strconv.Itoa(*semVersion.Patch)
-	return "(v" + majorVersion + "(\\." + minorVersion + "(\\." + patchVersion + ")?)?|" + versionRegexString + ")"
-}
-
-// GetMinorVersionRangeRegex generates minor version compatible range regex for the given version
-func GetMinorVersionRangeRegex(semVersion SemVersion) string {
-	versionRegexString := strings.ReplaceAll(semVersion.Version, ".", "\\.")
-	if semVersion.Patch == nil {
-		return "(v" + versionRegexString + "|" + versionRegexString + ")"
-	}
-	majorVersion := strconv.Itoa(semVersion.Major)
-	minorVersion := strconv.Itoa(semVersion.Minor)
-	patchVersion := strconv.Itoa(*semVersion.Patch)
-	return "(v" + majorVersion + "\\." + minorVersion + "(\\." + patchVersion + ")?|" + versionRegexString + ")"
-}
-
-// GetMajorVersionRange generates major version range for the given version
-func GetMajorVersionRange(semVersion SemVersion) string {
-	return "v" + strconv.Itoa(semVersion.Major)
-}
-
-// GetMinorVersionRange generates minor version range for the given version
-func GetMinorVersionRange(semVersion SemVersion) string {
-	return "v" + strconv.Itoa(semVersion.Major) + "." + strconv.Itoa(semVersion.Minor)
-}
-
-// CompareSemanticVersions compares two semantic versions and returns true
-// if `version` is greater or equal than `baseVersion`
-func CompareSemanticVersions(baseVersion, version SemVersion) bool {
-	if baseVersion.Major < version.Major {
-		return true
-	} else if baseVersion.Major > version.Major {
-		return false
-	} else {
-		if baseVersion.Minor < version.Minor {
-			return true
-		} else if baseVersion.Minor > version.Minor {
-			return false
-		} else {
-			if baseVersion.Patch != nil && version.Patch != nil {
-				if *baseVersion.Patch < *version.Patch {
-					return true
-				} else if *baseVersion.Patch > *version.Patch {
-					return false
-				}
-			} else if baseVersion.Patch == nil && version.Patch != nil {
-				return true
-			} else if baseVersion.Patch != nil && version.Patch == nil {
-				return false
-			}
-		}
-	}
-	return true
 }
